feat: refactoring read_annotations to module kubernetes_services
