+++ app/v1/modules/kubernetes_services/__init__.py
@@ -1,6 +1,7 @@
 # -*- coding: utf-8 -*-
 
 
+from .script_read_annotations import ScriptAnnotationsReaderservice
 from .script_to_dict import ScriptToDictParserService
 from .script_read_volumes import ScriptVolumesReaderService
 from .script_read_env import ScriptEnvironReaderService
@@ -9,6 +10,7 @@ from .script_read_volumes_mounts import ScriptVolumeMountsReaderService
 
 
 __all__ = [
+    "ScriptAnnotationsReaderservice",
     "ScriptToDictParserService",
     "ScriptVolumesReaderService",
     "ScriptEnvironReaderFromService",
